src/requirements_tf_260.txt
tf-agents==0.10.0
scipy==1.4.1
gym==0.24.1
matplotlib==3.3.4
numpy==1.19.5
tensorflow-probability==0.14.0
